Update Processing
Tried to fix some issues by drastically extending the training data. I think I will try with a slightly different model, as GaussianHMM doesn’t seem to be suiting the task well. I will try next with MultinomialHMM, which should help. The exact issue is that Gaussian assumes that the points exist on a scale, and that there are options inbetween the points, which isn’t really true (at least for this kind of music). The Multinomial requires tokenisation of the points first, and only deals in integers, so hopefully this will be fixed.