Update Processing
Okay, got it to work a bit better, by tokenising the input and snapping to the tokens. I feared that snapping to answers would be tricky, as the original input ended up will values such as 0.125, 0.1667, 0.0833, and many such values in between. I feared the inaccuracy of snapping in that circumstance, but now I feel it is a lot more fair to these small values, as the numeric interval remains the same between every token.
So to summarise the changes:
- data is now encoded as tokens before training
- the dataset now comprises of around 11,000 phrases
- the output is now snapped to the nearest real token in the dataset
- the phrase is then decoded and printed