First model
Just uploading the first model, I am not totally sure if it is a success, but it generates rhythm, so I guess it counts. It is trained on around 19,000 phrases reduced to just their rhythm. It was trained without barlines, or rests, so the aim is to create phrases that are based in classical teachings, but have a much higher degree of freedom.