Create Prompting.py
Finally got to the point where I can not only generate things, but also prompt the model. Next up will be mass testing. I’m planning on having one test where I prompt it 100 times with the same phrase starter, and then prompt it with 100 different example phrases out of the training data, to see if it can create new phrases that still adhere to the spirit of classical music.