Add files via upload
As for the naming protocol, g = gaussian, m1 or m2 regards the batch in which I trained them, s = states: s32 means 32 states, iter = the iteration setting for the model training. Essentially, the closer the number of states there are to the tokens in training data, the better it should be able to fit and predict the data. It also takes exponentially longer. The more iterations you allow, the more times the model will be able to go through the data and compare itself against it, so again, the more it will improve, at the huge time cost. 
With a gaussian HMM, there is a risk that when you run it for a long time, it might just fuck itself up, so I made these while testing the limits a bit. When I had the iter set to 100, it constantly returned an error after running for around 2 hours, which demolished my mental health. Doing it this way, meant I was able to produce batches of models, and I’ll be able to test them out to see how they compare to each other.